feat(block-publisher): support long block ranges and add GET endpoint
- switch block numbers and RPC ids from int to long across services and repositories
- add GET /v1/block/get endpoint to fetch blocks from a start to latest
- refactor BlockResource to reuse common range logic
- update scanner job and storage logic to use long ranges
- simplify BlockService by removing unused missing-block logic
- update docker-compose to build image locally instead of using prebuilt image

+++ src/main/java/com/simorghsoftech/api/requests/RpcRequest.java
@@ -2,7 +2,7 @@ package com.simorghsoftech.api.requests;
 
 import java.util.List;
 
-public record RpcRequest(int id,
+public record RpcRequest(long id,
                          String jsonrpc,
                          String method,
                          List<Object> params) {
